% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\title{Deprecated functions in ideal}
\arguments{
\item{...}{Ignored arguments.}
}
\value{
All functions throw a warning, with a deprecation message pointing
towards its descendent (if available).
}
\description{
Functions that are on their way to the function afterlife.
Their successors are also listed.
}
\details{
The successors of these functions are likely coming after the rework that
led to the creation of the \code{mosdef} package. See more into its
documentation for more details.
}
\section{Transitioning to the mosdef framework}{

\itemize{
\item \code{\link[=goseqTable]{goseqTable()}} is now being replaced by the more flexible
\code{\link[mosdef:run_goseq]{mosdef::run_goseq()}} function (which is even faster)
\item \code{\link[=ggplotCounts]{ggplotCounts()}} is now being replaced by the more flexible, better
designed, and actually even more good looking \code{\link[mosdef:gene_plot]{mosdef::gene_plot()}}
function, with better default behavior and all.
\item \code{\link[=deseqresult2tbl]{deseqresult2tbl()}} and \code{\link[=deseqresult2DEgenes]{deseqresult2DEgenes()}} are now replaced by the
more flexible \code{\link[mosdef:deresult_to_df]{mosdef::deresult_to_df()}}
\item The internally defined functions \code{createLinkENS()}, \code{createLinkGeneSymbol()},
and \code{createLinkGO()} are now replaced by the equivalent functions in \code{mosdef}:
\code{\link[mosdef:create_link_ENSEMBL]{mosdef::create_link_ENSEMBL()}}, \code{\link[mosdef:create_link_NCBI]{mosdef::create_link_NCBI()}} and
\code{\link[mosdef:create_link_GO]{mosdef::create_link_GO()}}. Notably, the \code{mosdef} package expanded on the
concept of automatically generated buttons, taking this to the extreme of
efficiency with the \code{\link[mosdef:buttonifier]{mosdef::buttonifier()}} function
}
}

\examples{
# try(goseqtable())

}
\author{
Federico Marini
}
