% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-pkg.R
\docType{package}
\name{ideal-pkg}
\alias{ideal-pkg}
\title{ideal: Interactive Differential Expression Analysis}
\description{
ideal makes differential expression analysis interactive, easy and reproducible.
The analysis of RNA-seq datasets is guided by the Shiny app as main component of
the package, which also provides a wide set of functions to efficiently extract
information from the existing data. The app can be also deployed on a Shiny
server, to allow its usage without any installation on the user's side.
}
\details{
ideal makes differential expression analysis interactive, easy and reproducible.
The analysis of RNA-seq datasets is guided by the Shiny app as main component of
the package, which also provides a wide set of functions to efficiently extract
information from the existing data. The app can be also deployed on a Shiny
server, to allow its usage without any installation on the user's side.
}
\author{
Federico Marini \email{marinif@uni-mainz.de}, 2016-2017

Maintainer: Federico Marini \email{marinif@uni-mainz.de}
}
