% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesignatures.R
\name{read_gmt}
\alias{read_gmt}
\title{Read in a GMT file}
\usage{
read_gmt(gmtfile)
}
\arguments{
\item{gmtfile}{A character value, containing the location of the GMT formatted
file. It can also be a file found online}
}
\value{
A list of vectors, one for each pathway in the GMT file.
}
\description{
Returns a list of pathways from a GMT file.
}
\examples{
# this example reads in the freely available pathways from wikipathways
mysigs <- read_gmt(
  "http://data.wikipathways.org/current/gmt/wikipathways-20180910-gmt-Homo_sapiens.gmt")
head(mysigs)
# see how the gene identifiers are encoded as ENTREZ id
}
