% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal_snapshot_oct14.R
\name{ideal}
\alias{ideal}
\title{Title}
\usage{
ideal(res_obj = NULL, dds_obj = NULL, annotation_obj = NULL,
  countmatrix = NULL, expdesign = NULL)
}
\arguments{
\item{annotation}{}
}
\description{
Title
}
\examples{




}

